+++ tests/unit/tickstore/test_toplevel.py
@@ -9,7 +9,7 @@ import pandas as pd
 import pytest
 from dateutil.rrule import rrule, DAILY
 from mockextras import when
-from pandas.util.testing import assert_frame_equal
+from pandas.testing import assert_frame_equal
 
 from arctic.date import DateRange, mktz
 from arctic.exceptions import OverlappingDataException