+++ tests/integration/chunkstore/tools/test_tools.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pandas as pd
-from pandas.util.testing import assert_frame_equal
+from pandas.testing import assert_frame_equal
 
 from arctic.chunkstore.tools import segment_id_repair
 from ..test_fixes import assert_frame_equal_