+++ tests/integration/store/test_version_store.py
@@ -10,7 +10,7 @@ import pandas as pd
 import pymongo
 import pytest
 from mock import Mock, patch
-from pandas.util.testing import assert_frame_equal, assert_series_equal
+from pandas.testing import assert_frame_equal, assert_series_equal
 from pymongo.errors import OperationFailure
 from pymongo.server_type import SERVER_TYPE