+++ tests/integration/tickstore/test_ts_read.py
@@ -5,9 +5,9 @@ import numpy as np
 import pandas as pd
 import pytest
 from mock import patch, call, Mock
-from numpy.testing.utils import assert_array_equal
+from numpy.testing import assert_array_equal
 from pandas import DatetimeIndex
-from pandas.util.testing import assert_frame_equal
+from pandas.testing import assert_frame_equal
 from pymongo import ReadPreference
 
 from arctic._util import mongo_count