+++ tests/unit/store/test_bitemporal_store.py
@@ -1,7 +1,7 @@
 from datetime import datetime as dt
 
 from mock import create_autospec
-from pandas.util.testing import assert_frame_equal
+from pandas.testing import assert_frame_equal
 
 from arctic.store.bitemporal_store import BitemporalStore
 from tests.util import read_str_as_pandas