+++ CHANGES.md
@@ -5,6 +5,7 @@
  * Compatibility: allow numpy version newer than 1.18
  * Build: specify which version of python/numpy/pandas to test
  * Bugfix: resolve import error on pandas.Panel, Panel was removed in pandas 1.0
+ * Bugfix: fix pandas get_timezone() argument cannot be None
  * Bugfix: fix np.float/np.int/np.book depreciations
  * Bugfix: fix numpy tostring() is deprecated. Use tobytes() instead.
  * Bugfix: numpy.testing.utils is deprecated