+++ CHANGES.md
@@ -4,7 +4,7 @@
  * Compatibility: update pinning from pandas<1.1, to pandas<2
  * Compatibility: update pinning from numpy<1.19 to numpy<2
  * Build: specify which version of python/numpy/pandas to test
- * Bugfix: resolve import error on pandas.Panel, Panel was removed in pandas 1.0
+ * Bugfix: #887 #928 resolve import error on pandas.Panel, Panel was removed in pandas 1.0
  * Bugfix: fix pandas get_timezone() argument cannot be None
  * Bugfix: fix np.float/np.int/np.book depreciations
  * Bugfix: fix numpy tostring() is deprecated. Use tobytes() instead.